year3/HCI/asl/requirements.txt
algorithms==0.1.4
appdirs==1.4.4
argon2-cffi==21.1.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beartype==0.11.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.4
Brotli==1.0.9
cached-property==1.5.2
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
cycler==0.11.0
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.7.0
Django==4.1.3
django-cors-headers==3.13.0
django-rest-framework==0.1.0
django-rest-framework-nested==0.0.1
djangorestframework==3.14.0
dnspython==2.1.0
docopt==0.6.2
entrypoints==0.4
fasteners==0.14.1
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
html5lib==1.1
httplib2==0.20.2
importlib-metadata==4.6.4
ioh==0.3.4
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==2.1.2
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
meson==0.61.2
monotonic==1.6
more-itertools==8.10.0
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
notebook==6.4.8
nptyping==2.3.1
numpy==1.23.4
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipreqs==0.4.11
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
Pygments==2.11.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-dateutil==2.8.1
python-dotenv==0.19.2
python-gnupg==0.4.8
python-xlib==0.32
python3-xlib==0.15
pythondialog==3.5.1
pythran==0.10.0
pytz==2022.1
PyUserInput==0.1.10
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
redis==4.3.4
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
scikit-learn==1.1.3
scipy==1.9.3
SecretStorage==3.3.1
Send2Trash==1.8.1b0
simplejson==3.17.6
six==1.16.0
sklearn==0.0.post1
soupsieve==2.3.1
sqlparse==0.4.3
sympy==1.9
terminado==0.13.1
testpath==0.5.0
texttable==1.6.4
threadpoolctl==3.1.0
tornado==6.1
traitlets==5.1.1
ufoLib2==0.13.1
unicodedata2==14.0.0
unix==1.0
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.2.2
widgetsnbextension==2.0.0
wrapt==1.14.1
xdg==5
yarg==0.1.9
zipp==1.0.0
